Fix dependency issue with hardhat tasks
